refactor bigint utility functions to test/utils.ts
